fixed input files

+++ src/lib/config/Config.cpp
@@ -38,7 +38,7 @@ void Config::print_usage() {
  * @return std::shared_ptr<Config> containing the configured parameters.
  *
  */
-auto Config::store_config_values(simulator::physics::ForceModel simulationType, BodyType bodyType, std::string &baseName, double endTime, double outputFrequency,
+auto Config::store_config_values(simulator::physics::ForceModel simulationType, BodyType bodyType, std::string &baseName, double endTime, long outputFrequency,
                                  std::string &outputFilename, int totalBodies, double deltaT, double inputSigma, double inputEpsilon, double massM, double distanceH,
                                  double averageBrownianMotion, std::vector<CelestialBody> celestialBodies, std::vector<Cuboid> cuboidVector,
                                  std::vector<Disc> discVector, int RngSeed) -> std::shared_ptr<Config> {
